python/google-agentic-stack/pyproject.toml
[project]
name = "google-agentic-stack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk[a2a,eval]>=1.19.0",
]
